refactor(*): move managers to main folder
