+++ src/application/ApplicationRequestManager.js
