+++ src/application/NestEggsManager.js
