+++ src/application/NodeLocationManager.js
