+++ src/application/NodeManager.js
