+++ src/application/UserManager.js
